+++ CieotApplication/src/main/java/com/lima/service/ITypeService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.lima.dto.TypeDTO;
 
 public interface ITypeService {
-	void setType(Long accountId, Integer typeId);
+	void setType(Integer accountId, Integer typeId);
 
 	List<TypeDTO> getAllType();
 }
